Adjust distance measures.

+++ prefrec-stratification/src/main/java/br/ufu/facom/lsi/prefrec/stratification/model/StratifiedMatrix.java
@@ -12,64 +12,64 @@ public class StratifiedMatrix implements Serializable {
 
 	private Long[] usersId;
 	private Long[] itemsId;
-	private Integer[][] ratings;
+	private Double[][] ratings;
 
-	private TreeMap<Integer, List<Integer[]>> partitions;
-	private TreeMap<Integer, List<Integer[]>> partitionsFromItem;
+	private TreeMap<Integer, List<Double[]>> partitions;
+	private TreeMap<Integer, List<Double[]>> partitionsFromItem;
 
 	private int columnIdx;
 	private int rowIdx;
 
 	public StratifiedMatrix(int folds, int rowLenght, int columnLenght) {
-		this.partitions = new TreeMap<Integer, List<Integer[]>>();
-		this.ratings = new Integer[rowLenght][columnLenght];
+		this.partitions = new TreeMap<Integer, List<Double[]>>();
+		this.ratings = new Double[rowLenght][columnLenght];
 		this.itemsId = new Long[columnLenght];
 		this.usersId = new Long[rowLenght];
 	}
 
-	public void addColumn(Integer[] col, int stratum, int itemId, Long[] users) {
+	public void addColumn(Double[] col, int stratum, int itemId, Long[] users) {
 		this.usersId = users;
 		this.itemsId[columnIdx] = Long.valueOf(itemId);
 		for (int i = 0; i < col.length; i++) {
 			this.ratings[i][this.columnIdx] = col[i];
 		}
 		this.columnIdx++;
-		List<Integer[]> temp = this.partitions.get(stratum);
+		List<Double[]> temp = this.partitions.get(stratum);
 		if (temp != null) {
 			temp.add(col);
 		} else {
-			temp = new ArrayList<Integer[]>();
+			temp = new ArrayList<Double[]>();
 			temp.add(col);
 			this.partitions.put(stratum, temp);
 		}
 	}
 
-	public void addRow(Integer[] row, int stratum, int userId, Long[] itens) {
+	public void addRow(Double[] row, int stratum, int userId, Long[] itens) {
 		this.itemsId = itens;
 		this.usersId[rowIdx] = Long.valueOf(userId);
 		for (int i = 0; i < row.length; i++) {
 			this.ratings[this.rowIdx][i] = row[i];
 		}
 		this.rowIdx++;
-		List<Integer[]> temp = this.partitions.get(stratum);
+		List<Double[]> temp = this.partitions.get(stratum);
 		if (temp != null) {
 			temp.add(row);
 		} else {
-			temp = new ArrayList<Integer[]>();
+			temp = new ArrayList<Double[]>();
 			temp.add(row);
 			this.partitions.put(stratum, temp);
 		}
 	}
 
-	public void addRestOfRows(Map<Integer, Integer[]> restOfRows) {
-		for (Map.Entry<Integer, Integer[]> entry : restOfRows.entrySet()) {
+	public void addRestOfRows(Map<Integer, Double[]> restOfRows) {
+		for (Map.Entry<Integer, Double[]> entry : restOfRows.entrySet()) {
 			this.addRow(entry.getValue(), this.partitions.lastKey(),
 					entry.getKey(), this.itemsId);
 		}
 	}
 
-	public void addRestOfColumns(Map<Integer, Integer[]> restOfColumns) {
-		for (Map.Entry<Integer, Integer[]> entry : restOfColumns.entrySet()) {
+	public void addRestOfColumns(Map<Integer, Double[]> restOfColumns) {
+		for (Map.Entry<Integer, Double[]> entry : restOfColumns.entrySet()) {
 			this.addColumn(entry.getValue(), this.partitions.lastKey(),
 					entry.getKey(), this.usersId);
 		}
@@ -78,14 +78,14 @@ public class StratifiedMatrix implements Serializable {
 	/**
 	 * @return the ratings
 	 */
-	public Integer[][] getRatings() {
+	public Double[][] getRatings() {
 		return ratings;
 	}
 
 	/**
 	 * @return the partitions
 	 */
-	public TreeMap<Integer, List<Integer[]>> getPartitions() {
+	public TreeMap<Integer, List<Double[]>> getPartitions() {
 		return partitions;
 	}
 
@@ -106,7 +106,7 @@ public class StratifiedMatrix implements Serializable {
 	/**
 	 * @return the partitionsFromItem
 	 */
-	public TreeMap<Integer, List<Integer[]>> getPartitionsFromItem() {
+	public TreeMap<Integer, List<Double[]>> getPartitionsFromItem() {
 		return partitionsFromItem;
 	}
 
@@ -115,7 +115,7 @@ public class StratifiedMatrix implements Serializable {
 	 *            the partitionsFromItem to set
 	 */
 	public void setPartitionsFromItem(
-			TreeMap<Integer, List<Integer[]>> partitionsFromItem) {
+			TreeMap<Integer, List<Double[]>> partitionsFromItem) {
 		this.partitionsFromItem = partitionsFromItem;
 	}
